+++ src/main/java/pupkin/brewersdelight/misc/BeverageProperties.java
@@ -3,134 +3,102 @@ package pupkin.brewersdelight.misc;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.food.FoodProperties;
+import vectorwing.farmersdelight.common.registry.ModEffects;
 
 public class BeverageProperties
-{   // TODO: Set unique effects for every beverage.
+{
 	// Farmers Delight
 	public static final FoodProperties BRAGA = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_BOOST, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties BRANDY = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.WATER_BREATHING, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties CIDER = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties GIN = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties KVASS = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
 			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties LIQUEUR = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties MARTINI = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(ModEffects.COMFORT.get(), 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties MELON_SCHNAPPS = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.BLINDNESS, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_BOOST, 1600, 2), 1.0f)
 			.build();
 
 	public static final FoodProperties SAKE = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.ABSORPTION, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties SBITEN = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
 			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties TEQUILA = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties WHISKY = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_BOOST, 1600, 2), 1.0f)
 			.build();
 
-	// Corn Delight
+	// Corn
 	public static final FoodProperties BOURBON = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 1600, 1), 1.0f)
 			.build();
 	public static final FoodProperties CORN_WHISKY = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 1600, 1), 1.0f)
 			.build();
 	public static final FoodProperties MOONSHINE = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.HEALTH_BOOST, 1600, 1), 1.0f)
 			.build();
 
-	// Italian Delight
+	// Grapes (Italian Delight)
 	public static final FoodProperties CAHORS = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.INVISIBILITY, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties CHAMPAGNE = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.JUMP, 1600, 1), 1.0f)
 			.build();
 
 	public static final FoodProperties COGNAC = new FoodProperties.Builder()
-			.nutrition(0)
-			.saturationMod(0.5f)
 			.alwaysEat()
-			.effect(() -> new MobEffectInstance(MobEffects.GLOWING, 1600, 1), 1.0f)
+			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 1600, 1), 1.0f)
 			.build();
-
 }